代码排雷整改

+++ services/native/src/usb_right_manager.cpp
@@ -129,6 +129,10 @@ bool UsbRightManager::HasRight(const std::string &deviceName, const std::string 
     std::shared_ptr<UsbRightDbHelper> helper = UsbRightDbHelper::GetInstance();
     // no record or expired record: expired true, has right false, add right next time
     // valid record: expired false, has right true, no need add right
+    if (helper == nullptr) {
+        USB_HILOGW(MODULE_USB_SERVICE, "helper is nullptr, false");
+        return false;
+    }
     return !helper->IsRecordExpired(userId, deviceName, bundleName, tokenId, nowTime);
 }
 
@@ -180,6 +184,10 @@ bool UsbRightManager::AddDeviceRight(const std::string &deviceName, const std::s
     info.validPeriod = USB_RIGHT_VALID_PERIOD_SET;
 
     std::shared_ptr<UsbRightDbHelper> helper = UsbRightDbHelper::GetInstance();
+    if (helper == nullptr) {
+        USB_HILOGW(MODULE_USB_SERVICE, "helper is nullptr, false");
+        return false;
+    }
     ret = helper->AddOrUpdateRightRecord(uid, deviceName, hapTokenInfoRes.bundleName, tokenIdStr, info);
     if (ret < 0) {
         USB_HILOGE(MODULE_USB_SERVICE, "add or update failed: %{public}s/%{public}d, ret=%{public}d",
@@ -211,6 +219,10 @@ bool UsbRightManager::AddDeviceRight(const std::string &deviceName, const std::s
     info.validPeriod = USB_RIGHT_VALID_PERIOD_SET;
 
     std::shared_ptr<UsbRightDbHelper> helper = UsbRightDbHelper::GetInstance();
+    if (helper == nullptr) {
+        USB_HILOGW(MODULE_USB_SERVICE, "helper is nullptr, false");
+        return false;
+    }
     auto ret = helper->AddOrUpdateRightRecord(userId, deviceName, bundleName, tokenId, info);
     if (ret < 0) {
         USB_HILOGE(MODULE_USB_SERVICE, "add or update failed: %{public}s/%{public}s/%{public}d, ret=%{public}d",
@@ -228,6 +240,10 @@ bool UsbRightManager::RemoveDeviceRight(const std::string &deviceName, const std
         return true;
     }
     std::shared_ptr<UsbRightDbHelper> helper = UsbRightDbHelper::GetInstance();
+    if (helper == nullptr) {
+        USB_HILOGW(MODULE_USB_SERVICE, "helper is nullptr, false");
+        return false;
+    }
     int32_t ret = helper->DeleteRightRecord(userId, deviceName, bundleName, tokenId);
     if (ret < 0) {
         USB_HILOGE(MODULE_USB_SERVICE, "delete failed: %{public}s/%{public}s/%{public}d", deviceName.c_str(),
@@ -273,6 +289,10 @@ bool UsbRightManager::ShowUsbDialog(
     want.SetParam("productName", productName);
 
     sptr<UsbAbilityConn> usbAbilityConn_ = new (std::nothrow) UsbAbilityConn();
+    if (usbAbilityConn_ == nullptr) {
+        USB_HILOGW(MODULE_USB_SERVICE, "usbAbilityConn_ is nullptr, false");
+        return false;
+    }
     sem_init(&waitDialogDisappear_, 1, 0);
     auto ret = abmc->ConnectAbility(want, usbAbilityConn_, -1);
     if (ret != UEC_OK) {
@@ -497,6 +517,10 @@ int32_t UsbRightManager::CleanUpRightAppUninstalled(int32_t uid, int32_t &totalA
 {
     std::vector<std::string> apps;
     std::shared_ptr<UsbRightDbHelper> helper = UsbRightDbHelper::GetInstance();
+    if (helper == nullptr) {
+        USB_HILOGW(MODULE_USB_SERVICE, "helper is nullptr, false");
+        return false;
+    }
     int32_t ret = helper->QueryRightRecordApps(uid, apps);
     if (ret <= 0) {
         /* error or empty record */
@@ -506,7 +530,7 @@ int32_t UsbRightManager::CleanUpRightAppUninstalled(int32_t uid, int32_t &totalA
     deleteApps = 0;
     for (int32_t i = 0; i < totalApps; i++) {
         std::string app = apps.at(i);
-        if (!IsAppInstalled(uid, app)) {
+        if (helper != nullptr && (!IsAppInstalled(uid, app))) {
             ret = helper->DeleteAppRightRecord(uid, app);
             if (ret != USB_RIGHT_OK) {
                 USB_HILOGW(MODULE_USB_SERVICE, "clean failed: app=%{public}s, ret=%{public}d", app.c_str(), ret);
@@ -524,6 +548,10 @@ int32_t UsbRightManager::CleanUpRightAppUninstalled(int32_t uid, const std::stri
 {
     std::vector<std::string> apps;
     std::shared_ptr<UsbRightDbHelper> helper = UsbRightDbHelper::GetInstance();
+    if (helper == nullptr) {
+        USB_HILOGW(MODULE_USB_SERVICE, "helper is nullptr, false");
+        return false;
+    }
     int32_t ret = helper->QueryRightRecordApps(uid, apps);
     if (ret <= 0) {
         /* error or empty record */
